Updated CPO to be compliant with JDBC 4.3 (#40)
* Updated CPO to be compliant with JDBC 4.3
- Added new JDBC Types
- Added Java Types that JDBC will convert using setObject(index) and
  getObject(index, Class<T>)
- Modified the ResultSet getters, PreparedStatementSetters and
  CallStatement getters and setters to use the new setObject and getObject facilities
- Added a new method mapper type for the setObject and getObject
  facilities
- Changed the exporter code to use the javaClass and not the dataClass
  as they can be different now.
- Added a test to test the setObject and getObject

* added in one more assert to make sure the setObject and getObject type
conversions are working

+++ cpo-jdbc/src/test/java/org/synchronoss/cpo/jdbc/ValueObject.java
@@ -21,6 +21,7 @@
 package org.synchronoss.cpo.jdbc;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.sql.Date;
 import java.sql.Time;
 import java.sql.Timestamp;
@@ -72,7 +73,7 @@ public interface ValueObject {
 
   public int getAttrSmallInt();
 
-  public BigDecimal getAttrBigInt();
+  public BigInteger getAttrBigInt();
 
   public BigDecimal getAttrReal();
 
@@ -122,7 +123,7 @@ public interface ValueObject {
 
   public void setAttrSmallInt(int attrSmallInt);
 
-  public void setAttrBigInt(BigDecimal attrBigInt);
+  public void setAttrBigInt(BigInteger attrBigInt);
 
   public void setAttrReal(BigDecimal attrReal);
